Fixed the HUGE STUPID bug in the gePossibleOthers where it seemed like the data become garbage
- Eventually, it was just a bug in the code, but it lead to weird outcomes. I think that the compiler optimized
the code in an unexpected way because of the bug. so for example, i had there a bool that is always false, this shadowed some logic
where i use some variable. It is highly likely that the compiler or runtime just optimized out the variable leading me to see it
as garbage because it was never needed....